N00BS: fix issue where process would keep running in the background when closed via the X button

+++ HorseIsleServer/MPN00BS/ServerStarter.cs
@@ -52,7 +52,8 @@ namespace MPN00BS
 
         public static void CloseHorseIsleClient()
         {
-            clientProcess.Kill();
+            if(!clientProcess.HasExited)
+                clientProcess.Kill();
         }
         public static void StartHorseIsleClient(Action callback, string serverIp, int serverPort)
         {
